# lb-rs: perf inquiry ch 1