--- a/libs/lb/lb-rs/src/model/errors.rs
+++ b/libs/lb/lb-rs/src/model/errors.rs
@@ -502,6 +502,9 @@ impl From<SharedError> for TestRepoError {
 
 impl From<LbErr> for TestRepoError {
     fn from(value: LbErr) -> Self {
+        if value.kind == LbErrKind::AccountNonexistent {
+            return Self::NoAccount;
+        }
         Self::Core(value)
     }
 }