… Cargo.toml
… libs/lb/lb-rs/src/logic/core_ops.rs
… libs/lb/lb-rs/src/logic/staged.rs
… libs/lb/lb-rs/src/logic/tree_like.rs
… libs/lb/lb-rs/src/logic/validate.rs
… libs/lb/lb-rs/src/model/core_config.rs
… libs/lb/lb-rs/src/model/file_metadata.rs
… libs/lb/lb-rs/src/service/account.rs
… libs/lb/lb-rs/src/service/documents.rs
… libs/lb/lb-rs/src/service/file.rs
… libs/lb/lb-rs/src/service/integrity.rs
… libs/lb/lb-rs/src/service/keychain.rs
… libs/lb/lb-rs/src/service/network.rs
… libs/lb/lb-rs/src/service/path.rs
… libs/lb/lb-rs/src/service/search.rs
… libs/lb/lb-rs/src/service/share.rs
… libs/lb/lb-rs/src/service/sync.rs
… libs/lb/lb-rs/src/service/usage.rs
… libs/lb/lb-rs/tests/account_service_tests.rs
… libs/lb/lb-rs/tests/api_new_account_tests.rs
… libs/lb/lb-rs/tests/cli_data_tests.rs
… libs/lb/lb-rs/tests/core_ops_tests.rs
… libs/lb/lb-rs/tests/integrity_tests.rs
… libs/lb/lb-rs/tests/lazy_tests.rs
… libs/lb/lb-rs/tests/path_tests.rs
… libs/lb/lb-rs/tests/sync_service_share_tests.rs
… libs/lb/lb-rs/tests/tree_composition_tests.rs
… libs/lb/lb-rs/tests/validate_tests.rs
… libs/lb/test_utils/src/assert.rs
… server/src/account_service.rs