@Parth 2024-12-24
It was also seen at points that people had 100ms+ values in their logs. The idle time here was low. I suspect that the tokio mutex doesn't show up properly in logs.

It would be easy to verify and I don't think our use of the tokio mutex is essential, but I'm also skeptical this is the case because the same people make both. so further investigation is certainly required.