@Parth 2024-12-27
Those several hundred millisecond log lines were misread by me -- they were list metas, which were genuinely taking that long, and are now generally less than 1ms (500+ x speedup).